PEP8 and other code cleaning.
