--- mission_pe/MyPage.html
@@ -1,77 +1,0 @@
-
-{% block content %}
-    <p><strong>
-        Please rank the following charities from 1 to 4 in order of your true preference (1 being most preferred charity
-        and 4 being the charity that you do not support). Your choice will not impact your payment. In the experiment, your decisions can
-        generate a monetary donation for the charity of your choice, in addition to you own payment. Remember, data
-        collected is anonymised.
-    </strong></p>
-
-    {% block form_content %}
-
-            <div class="container border bg-light p-2 mt-2 mb-2">
-                <div class="mb-3">
-                  <p>
-                    <strong>CISWO (Coal Industry Social Welfare Organisation) - </strong>
-                    “We’re a UK-wide charity providing free support to former coal miners and their families….We also work to
-                    secure the future of recreational facilities in former coal mining communities and work with local mining
-                    charities to ensure their services continue.”
-                    {% formfield 'CISWO' %}
-                  </p>
-
-                </div>
-            </div>
-
-            <div class="container border bg-light p-2 mt-2 mb-2">
-                <div class="mb-3">
-                    <p>
-                        <strong>The Crowd: Ember - </strong>
-                        “Funds paid to Ember are used in their research and policy campaigning against the continued use of coal….Ember
-                        provide expert advice for campaigning organisations across the world, ensuring the right pressure is applied, at
-                        the right time, to close coal power plants."
-                    </p>
-                    {% formfield 'Ember' %}
-
-                </div>
-            </div>
-
-            <div class="container border bg-light p-2 mt-2 mb-2">
-                <div class="mb-3">
-                    <p>
-                        <strong>CARE (Christian Action, Research, and Education) - </strong>
-                        “CARE is passionately pro-woman and pro-life: we advocate for both lives in a pregnancy to be protected and for
-                        a society where women are genuinely supported….We want to work for a society where abortion is unthinkable, and
-                        women and babies are protected.”
-                    </p>
-                    {% formfield 'CARE' %}
-
-                </div>
-            </div>
-
-            <div class="container border bg-light p-2 mt-2 mb-2">
-                <div class="mb-3">
-                    <p>
-                        <strong>BPAS (British Pregnancy Advisory Service) - </strong>
-                        “Purpose: To remove all barriers to reproductive choice and to advocate for and deliver high quality,
-                        woman-centred reproductive health care.…. Over the years we have worked with the National Health Service to
-                        ensure that access to abortion care is equitable, fast and free for the client.”
-                    </p>
-                    {% formfield 'BPAS' %}
-                </div>
-            </div>
-
-    {% endblock form_content %}
-
-        {{ next_button }}
-
-
-    <footer>
-            <hr/>
-            <div class="text-right">
-                <p class="text-muted font-italic">Laboratory for Economic and Decision Research, University of East Anglia
-                    <img src="{% static 'global/uea.png' %}" class="align-content-center ml-3" alt="UEA logo"/>
-                </p>
-            </div>
-    </footer>
-
-{% endblock %}
