+++ payment_info/__init__.py
@@ -1,7 +1,6 @@
 from otree.api import *
 
 
-
 class Constants(BaseConstants):
     name_in_url = 'payment_info'
     players_per_group = None
@@ -20,23 +19,26 @@ class Player(BasePlayer):
     app_to_pay = models.StringField
     charity_to_pay = models.StringField
 
+
 # FUNCTIONS
 # PAGES
+
+
 class PaymentInfo(Page):
     @staticmethod
     def vars_for_template(player: Player):
         import random
         participant = player.participant
-        apps = ['stage1_final', 'stage2_2test', 'stage3']
+        apps = ['stage1', 'stage2', 'stage3']
         app_to_pay = random.choice(apps)
         participant.payoff = participant.app_payoffs[app_to_pay]
         player.app_to_pay = app_to_pay
 
-        charity = ['stage2_2test', 'stage3']
+        charity = ['stage2', 'stage3']
         charity_to_pay = random.choice(charity)
         player.charity_to_pay = charity_to_pay
 
 
-
-
-page_sequence = [PaymentInfo]
+page_sequence = [
+    PaymentInfo
+]
