+++ stage1/Instructions.html
@@ -1,0 +1,41 @@
+{% block title %}
+    Stage 1 : Instructions
+{% endblock %}
+{% block content %}
+    <p>
+        For each round, your screen will show 48 randomly positioned sliders. To correctly
+        complete a slider, it must be moved <em>exactly to the center marked by a red dot.</em> The slider will turn <em>green
+        when correctly completed.</em>
+        The screen will show you the timer and a progress bar indicating how many sliders you have completed. The next
+        round will begin automatically after 90 seconds with a new set of sliders.
+    </p>
+
+    <p>
+        You must move <em>as many sliders to the center as you can</em>. Sliders completed in all rounds will be counted.
+    </p>
+
+    <p><strong>
+        If this stage is chosen for your payment, you will receive £1.5 per round.
+    </strong></p>
+
+    <p>
+        Please raise your hand if you have any questions.
+    </p>
+
+    <p>
+        Click on 'next' to begin the first round.
+    </p>
+
+
+    {% next_button %}
+
+    <footer>
+            <hr/>
+            <div class="text-right">
+                <p class="text-muted font-italic">Laboratory for Economic and Decision Research, University of East Anglia
+                    <img src="{% static 'global/uea.png' %}" class="align-content-center ml-3" alt="UEA logo"/>
+                </p>
+        </div>
+        </footer>
+
+{% endblock %}
