+++ stage1/static/sliders.css
