+++ stage1/static/sliders.js
