+++ stage2/Termination.html
@@ -1,0 +1,57 @@
+{% block title %}
+Result for Stage 2
+{% endblock %}
+
+{% block content %}
+
+
+<p>
+    Your stage was terminated early as you completed fewer than 20 sliders in the previous round i.e. {% player.round_number %}.
+</p>
+
+<p>
+
+    {% if player.round_number == 1 %}
+    Sliders completed in round 1 = {% num_correct_s2_1 %}<br>
+
+    {% elif player.round_number == 2 %}
+    Sliders completed in round 1 = {% num_correct_s2_1 %}<br>
+    Sliders completed in round 2 = {% num_correct_s2_2 %}<br>
+
+    {% elif player.round_number == 3 %}
+    Sliders completed in round 1 = {% num_correct_s2_1 %}<br>
+    Sliders completed in round 2 = {% num_correct_s2_2 %}<br>
+    Sliders completed in round 3 = {% num_correct_s2_3 %}<br>
+
+    {% elif player.round_number == 4 %}
+    Sliders completed in round 1 = {% num_correct_s2_1 %}<br>
+    Sliders completed in round 2 = {% num_correct_s2_2 %}<br>
+    Sliders completed in round 3 = {% num_correct_s2_3 %}<br>
+    Sliders completed in round 4 = {% num_correct_s2_4 %}<br>
+
+    {% elif player.round_number == 5 %}
+    Sliders completed in round 1 = {% num_correct_s2_1 %}<br>
+    Sliders completed in round 2 = {% num_correct_s2_2 %}<br>
+    Sliders completed in round 3 = {% num_correct_s2_3 %}<br>
+    Sliders completed in round 4 = {% num_correct_s2_4 %}<br>
+    Sliders completed in round 5 = {% num_correct_s2_5 %}<br>
+
+    {% endif %}
+</p>
+
+<p>
+    <strong>If this stage is chosen for payment, your payoff will be = {% player.potential_payoff %}</strong>
+</p>
+
+{% next_button %}
+
+    <footer>
+            <hr/>
+            <div class="text-right">
+                <p class="text-muted font-italic">Laboratory for Economic and Decision Research, University of East Anglia
+                    <img src="{% static 'global/uea.png' %}" class="align-content-center ml-3" alt="UEA logo"/>
+                </p>
+        </div>
+        </footer>
+
+{% endblock %}
