+++ stage2/static/sliders.css
