+++ stage2/static/sliders.js
