--- stage2_2test/Game.html
@@ -1,41 +1,0 @@
-{% block styles  %}
-    <link href="{% static 'sliders.css' %}" rel="stylesheet"/>
-{% endblock %}
-
-{% block title %}
-Stage 2: Round {% player.round_number %} of 5
-{% endblock %}
-
-{% block content %}
-<div class="d-flex flex-column align-items-center">
-    <progress id="progress-bar" max="{% params.num_sliders %}" value="0"></progress>
-    <canvas id="canvas" class="m-2"></canvas>
-
-</div>
-<br>
-
-    <footer>
-            <hr/>
-            <div class="text-right">
-                <p class="text-muted font-italic">Laboratory for Economic and Decision Research, University of East Anglia
-                    <img src="{% static 'global/uea.png' %}" class="align-content-center ml-3" alt="UEA logo"/>
-                </p>
-        </div>
-        </footer>
-{% endblock %}
-
-{% block scripts %}
-<script src="{% static 'sliders.js' %}"></script>
-
-<script type="text/javascript">
-    history.pushState(null, null, location.href);
-    history.back();
-    history.forward();
-    window.onpopstate = function () { history.go(1); };
-</script>
-
-<SCRIPT type="text/javascript">
-	window.history.forward();
-	function noBack() { window.history.forward(); }
-</SCRIPT>
-{% endblock %}
